Just a umd and an esm script per source
