+++ docs/pages/pass.html
@@ -7,13 +7,16 @@
         <meta name="author" content="" />
         <title>JOOA11Y - The Joomla Accessibility Checker</title>
         <link rel="icon" type="image/x-icon" href="../assets/favicon.ico" />
-
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/joomla-projects/joomla-a11y-checker@joomla/dist/css/joomla-a11y-checker.min.css" />
-
+        <link rel="stylesheet" href="../assets/css/joomla-a11y-checker.min.css" />
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
             integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-            <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css"
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css"
             integrity="sha512-1ycn6IcaQQ40/MKBW2W4Rhis/DbILU74C1vSrLJxCq57o941Ym01SwNsOMqvEBFlcgUa6xLiPY/NS5R+E6ztJQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
+        <!-- Bootstrap core JS-->
+        <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js"
+          integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
+        <!-- Core theme JS-->
+        <script type="module" src="script.js"></script>
     </head>
         <body>
             <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
@@ -183,8 +186,5 @@
       ignored as it is inside the &lt;footer&gt;.</p><p> If a content editor can't edit it, then Jooa11y doesn't scan it.</p>
     </div>
 	</footer>
-		<!-- scripts -->
-		<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>        <!-- Core theme JS-->
-		<script type="module" src="script.js"></script>
 	</body>
 </html>
