+++ rollup.config.js
@@ -22,13 +22,13 @@ export default [
     })
     ],
     output: [
-      { file: 'dist/js/joomla-a11y-checker.standalone.esm.js', format: 'esm'},
+      { file: 'dist/js/joomla-a11y-checker.esm.js', format: 'esm'},
       {
-        file: 'dist/js/joomla-a11y-checker.standalone.esm.min.js', format: 'esm', plugins: [terser()],
+        file: 'dist/js/joomla-a11y-checker.esm.min.js', format: 'esm', plugins: [terser()],
       },
     ],
   },
-  // ES6 standalone files
+  // UMD standalone files
   {
     input: 'src/js/jooa11y.js',
     plugins: [
@@ -40,33 +40,9 @@ export default [
     })
     ],
     output: [
-      { file: 'dist/js/joomla-a11y-checker.standalone.umd.js', format: 'umd', name: 'Jooa11y'},
+      { file: 'dist/js/joomla-a11y-checker.umd.js', format: 'umd', name: 'Jooa11y'},
       {
-        file: 'dist/js/joomla-a11y-checker.standalone.umd.min.js', format: 'umd', name: 'Jooa11y', plugins: [terser()],
-      },
-    ],
-  },
-  // ES6 Joomla bundle files
-  {
-    input: 'src/js/jooa11y.js',
-    plugins: [
-      nodeResolve(),
-      css(),
-      replace({
-      preventAssignment: true,
-      'process.env.NODE_ENV': JSON.stringify('production'),
-      }),
-      // In the Joomla build tools, we need to use the Popper from Bootstrap:
-      // Replace import { applyStyles, createPopper } from '@popperjs/core';
-      // with import { applyStyles, createPopper } from '../../bootstrap/js/popper.js?v=x.x.x';
-    ],
-    external: [
-      '@popperjs/core'
-    ],
-    output: [
-      { file: 'dist/js/joomla-a11y-checker.esm.js', format: 'esm'},
-      {
-        file: 'dist/js/joomla-a11y-checker.esm.min.js', format: 'esm', plugins: [terser()],
+        file: 'dist/js/joomla-a11y-checker.umd.min.js', format: 'umd', name: 'Jooa11y', plugins: [terser()],
       },
     ],
   },
@@ -75,6 +51,13 @@ export default [
     plugins: [nodeResolve()],
     output: [
       { file: 'dist/js/lang/en.js', format: 'esm'},
+    ],
+  },
+  {
+    input: 'src/js/lang/en.js',
+    plugins: [nodeResolve()],
+    output: [
+      { file: 'dist/js/lang/en.umd.js', format: 'umd', name: 'Jooa11yLang'},
     ],
   },
   // SCSS files
